Splitting tests into different files
